no implicit type conversations, make sure we understand what goes in

+++ bliss/datasets/saved_datasets.py
@@ -19,8 +19,8 @@ class SavedGalsimBlends(Dataset):
         super().__init__()
         ds: dict[str, Tensor] = load_dataset_h5py(dataset_file)
 
-        self.images = ds.pop("images").float()  # needs to be a float for NN
-        self.background = ds.pop("background").float()
+        self.images = ds.pop("images")
+        self.background = ds.pop("background")
         self.epoch_size = len(self.images)
 
         # don't need for training
@@ -31,9 +31,9 @@ class SavedGalsimBlends(Dataset):
         # avoid large memory usage if we don't need padding.
         if not keep_padding:
             ds.pop("paddings")
-            self.paddings = torch.tensor([0]).float()
+            self.paddings = torch.tensor([0.0])
         else:
-            self.paddings = ds.pop("paddings").float()
+            self.paddings = ds.pop("paddings")
         self.keep_padding = keep_padding
 
         full_catalog = FullCatalog(slen, slen, ds)
@@ -58,8 +58,8 @@ class SavedIndividualGalaxies(Dataset):
         super().__init__()
         ds: dict[str, Tensor] = load_dataset_h5py(dataset_file)
 
-        self.images = ds.pop("images").float()  # needs to be a float for NN
-        self.background = ds.pop("background").float()
+        self.images = ds.pop("images")
+        self.background = ds.pop("background")
 
         self.epoch_size = len(self.images)
